Handle errors

+++ app/lib/database/review/review.mutation.ts
@@ -4,14 +4,15 @@ import { ProductModel, ReviewModel } from "@/app/lib/database/models";
 import { Product, Review } from "@/app/lib/definitions";
 
 export async function createReview(
-  userId: string,
+  name: string,
+  email: string,
   productId: string,
   rating: number,
   comment: string,
 ) {
   try {
-    const review = await ReviewModel.findOne({
-      user: userId,
+    const review = await ReviewModel.findOne<HydratedDocument<Review>>({
+      email: email,
       product: productId,
     });
 
@@ -22,7 +23,8 @@ export async function createReview(
 
     // Otherwise, create and save the review
     const newReview: HydratedDocument<Review> = new ReviewModel({
-      user: userId,
+      name,
+      email,
       rating,
       comment,
       product: productId,
@@ -32,12 +34,21 @@ export async function createReview(
 
     // Update product rating and numReviews;
     const productToReview = await ProductModel.findById(productId);
-    const prevRating = productToReview.rating;
-    const prevNumReview = productToReview.numReview;
+    const allReviews = await ReviewModel.find().lean().exec();
+    const numReviews = allReviews.length;
+    const ratingSum = allReviews.reduce((acc, review) => {
+      return acc + review.rating;
+    }, 0);
+    let productRating = 0;
+    if (numReviews > 0) {
+      productRating = ratingSum / numReviews;
+    }
 
-    productToReview.rating = (prevRating + rating) / (prevNumReview + 1);
-    productToReview.numReview = prevNumReview + 1;
-    await productToReview.save();
+    if (productToReview) {
+      productToReview.rating = productRating;
+      productToReview.numReviews = numReviews;
+      await productToReview.save();
+    }
 
     // Update cache
     revalidatePath(`/products/${productId}`);
@@ -46,27 +57,35 @@ export async function createReview(
   }
 }
 
-export async function deleteReview(id: string) {
+export async function deleteReview(reviewId: string) {
   try {
     // Find the review with this id
-    const reviewToDelete = await ReviewModel.findById(id).exec();
+    const reviewToDelete = await ReviewModel.findById(reviewId).exec();
+    const allReviews = await ReviewModel.find({ _id: { $ne: reviewId } })
+      .lean()
+      .exec();
+    const numReviews = allReviews.length;
+    const ratingSum = allReviews.reduce((acc, review) => {
+      return acc + review.rating;
+    }, 0);
+    let rating = 0;
+    if (numReviews > 0) {
+      rating = ratingSum / numReviews;
+    }
+    const productId = reviewToDelete.product;
 
     // Find the reviewed  and update rating and numReviews
-    const productReviewed = await ProductModel.findById<
-      HydratedDocument<Product>
-    >(reviewToDelete.product).exec();
+    const productReviewed =
+      await ProductModel.findById<HydratedDocument<Product>>(productId).exec();
 
-    if (productReviewed) {
-      const prevRating = productReviewed.rating;
-      const prevNumReview = productReviewed.numReviews;
-
-      productReviewed.rating =
-        (prevRating - reviewToDelete.rating) / (prevNumReview - 1);
-      productReviewed.numReviews = prevNumReview - 1;
+    if (productReviewed && reviewToDelete) {
+      productReviewed.rating = rating;
+      productReviewed.numReviews = numReviews;
 
       // Commit opertations
       await reviewToDelete.deleteOne();
       await productReviewed.save();
+      revalidatePath(`/admin/products/${productId}`);
     }
   } catch (err) {
     console.error(err);
@@ -76,6 +95,18 @@ export async function deleteReview(id: string) {
 export async function deleteProductReviews(productId: string) {
   try {
     await ReviewModel.deleteMany({ product: productId });
+
+    // Reset product rating
+    const productToDeleteReviews =
+      await ProductModel.findById(productId).exec();
+    if (productToDeleteReviews) {
+      productToDeleteReviews.rating = 0;
+      productToDeleteReviews.numReviews = 0;
+      productToDeleteReviews.save();
+
+      // Revalidate cache
+      revalidatePath(`/admin/products/${productId}`);
+    }
   } catch (err) {
     console.error(err);
   }
